Make package installable
