+++ .gitignore
@@ -1,0 +1,6 @@
+# python build artifacts
+__pycache__
+*.egg-info
+.ipynb_checkpoints
+build
+dist
