+++ setup.cfg
@@ -1,0 +1,34 @@
+
+[metadata]
+name = dash-bootstrap-templates
+description = A collection of Plotly figure templates with a Bootstrap theme
+long_description = file: README.md
+long_description_content_type = text/markdown
+license = MIT
+author = AnnMarieW
+# author_email =
+url = https://github.com/AnnMarieW/dash-bootstrap-templates
+classifiers =
+    Development Status :: 3 - Alpha
+    License :: OSI Approved :: MIT License
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+
+[options]
+package_dir=
+    =src
+packages=find:
+include_package_data = True
+install_requires =
+    dash
+    dash-bootstrap-components
+    importlib_metadata>=3.4.0,<4.0.0; python_version == "3.7"
+    importlib_resources>=5.1.0,<6.0.0; python_version < "3.9"
+    numpy
+
+[options.packages.find]
+where = src
+
+[options.package_data]
+dash_bootstrap_templates = templates/*.json
