+++ src/dash_bootstrap_templates/templates/cosmo.json
