+++ src/dash_bootstrap_templates/templates/flatly.json
