+++ src/dash_bootstrap_templates/templates/journal.json
