+++ src/dash_bootstrap_templates/templates/minty.json
