+++ src/dash_bootstrap_templates/templates/pulse.json
