+++ src/dash_bootstrap_templates/templates/simplex.json
