+++ src/dash_bootstrap_templates/templates/sketchy.json
