+++ src/dash_bootstrap_templates/templates/slate.json
