+++ src/dash_bootstrap_templates/templates/solar.json
