+++ src/dash_bootstrap_templates/templates/spacelab.json
