+++ src/dash_bootstrap_templates/templates/yeti.json
